% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_names.R
\docType{data}
\name{column_names}
\alias{column_names}
\title{List of OpenFace column names}
\format{
An object of class \code{character} of length 329.
}
\usage{
column_names
}
\arguments{
\item{column_names}{A character vector of column names from OpenFace output.}
}
\description{
List of OpenFace column names
}
\keyword{datasets}
