% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rows.R
\name{filter_rows}
\alias{filter_rows}
\title{Filter rows based on your exclusion criteria.}
\usage{
filter_rows(
  dat,
  conf_removal,
  succ_removal,
  conf_thres,
  succ_thres,
  output_dir
)
}
\arguments{
\item{dat}{Your long-form dataframe.}

\item{conf_removal}{Numeric. Do you want to remove entire participants based on average confidence score?
Specify threshold here. Will keep all participants above this average. Defaults to 0.}

\item{succ_removal}{Do you want to remove entire participants based on average number of unsuccessful rows?
Specify threshold here. Will keep all participants above this threshold. Defaults to 0.}

\item{conf_thres}{Numeric. Your confidence threshold. Will keep all rows above the specified threshold.}

\item{succ_thres}{Binary, TRUE/FALSE. Do you want to remove individual rows where success is zero? Defaults to FALSE.}

\item{output_dir}{A \code{.txt} file will generated on who was removed and what the exclusion criteria was.
Where should this output file be saved?}
}
\description{
Filter rows and/or remove entire participants with poor confidence and/or low success.
Note that \code{conf_removal} and \code{succ_removal} will occur before row filtering in that order.
Then rows will be filtered first by confidence, then success.
}
