% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_cleaner.R
\name{id_cleaner}
\alias{id_cleaner}
\title{Clean your participant identifier.}
\usage{
id_cleaner(dat, id_num)
}
\arguments{
\item{dat}{Dataframe}

\item{id_num}{Participant identifier.}
}
\description{
Removes prefix from your participant identifier.
This assumes your participant ID number are all the digits following the last underscore (e.g., "OpenFace_01234").
Do not apply to raw dyadic data (e.g., "OpenFace_01234_Mother").
}
