% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_stack.R
\name{long_stack}
\alias{long_stack}
\title{Merge/stack all dataframes in your list into one long dataframe.}
\usage{
long_stack(yourlist)
}
\arguments{
\item{yourlist}{List of dataframes, assuming it's already been loaded into R.}
}
\description{
This will take a ridiculously long time.
}
