% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rows.R, R/make_fake2.R
\name{make_fake2}
\alias{make_fake2}
\title{Filter rows based on your exclusion criteria.}
\usage{
make_fake2(path, n, loop, dyad, seed)

make_fake2(path, n, loop, dyad, seed)
}
\arguments{
\item{path}{Character, path to where you want the files to go.}

\item{n}{Numeric, number of rows for each generated \code{.csv}}

\item{loop}{Numeric, total number of individual \code{.csv} files to generate.}

\item{dyad}{Two dyad members as a character vector. E.g., \verb{c("Mother", "Daughter)}.}

\item{seed}{Numeric, random seed. Optional.}
}
\description{
Filter rows and/or remove entire participants with poor confidence and/or low success.
To use this function, you must already have one single dataset in long-form,
like after using \code{long_stack()}.

Each \code{.csv} file will have a random 5-digit ID number with the prefix "FakeOpenFace_".
Excluding the informational (e.g., \code{faceid}, \code{timestamp}) and the Action Unit data,
since the values don't make sense in this situation.
Each generated \code{.csv} file will end with a suffix indicating dyad member.
}
