% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_small.R
\name{make_small}
\alias{make_small}
\title{Make OpenFace \code{.csv} files narrower}
\usage{
make_small(bigpath, newpath)
}
\arguments{
\item{bigpath}{Path to where all your big files currently live.}

\item{newpath}{Path to where you want to write out the new smaller files.}
}
\description{
We only need the first few columns and the Action Unit columns any way.
}
