% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_video.R
\name{name_video}
\alias{name_video}
\title{Generate one line of text for FeatureExtraction.exe}
\usage{
name_video(input_dir, output_dir, of_dir, save_txt, specify)
}
\arguments{
\item{input_dir}{Directory of where your files are stored.}

\item{output_dir}{Directory of where you want the processed OpenFace videos to be saved.}

\item{of_dir}{Directory of where your command is stored. Must include \code{FeatureExtraction.exe} at the end.}

\item{save_txt}{Directory of where you want the .txt file to be saved.}

\item{specify}{Optional. Adds \code{-aus} to specify that only Action Units should be computed Defaults to \code{FALSE}.}
}
\description{
Generate one line of text to run in the Command Line for FeatureExtraction.exe
OpenFace does not allow you to run all the videos in a single directory,
but does allow multiple videos on the same line if each video is individually specified.
This may be able to be done with a Python tool, but this may be easier,
since I don't have to rely on a third party tool or download Python.
}
