% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_dyad.R
\name{pivot_dyad}
\alias{pivot_dyad}
\title{Widen data by dyad member.}
\usage{
pivot_dyad(dat, output_dir)
}
\arguments{
\item{dat}{Dataframe.}

\item{output_dir}{Where you want your dataframes to be saved to.}
}
\description{
Widen dataframe by dyad member.
At this point, you should have merged your data using \code{load_list} and \verb{long stack}
and filtered using relevant criteria using \code{filter_rows}.
This will output two dataframes: 1) an extra long dataframe by dyad member, and 2)
a slightly wider dataframe where each dyad members has their own Action Unit column.
}
