% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_stats.R
\name{print_stats}
\alias{print_stats}
\title{Calculate statistics on confidence and succcess for a dataframe.}
\usage{
print_stats(dat, id_num, output_dir)
}
\arguments{
\item{dat}{Dataframe}

\item{id_num}{Participant identifier variables. Defaults to the first column in your dataframe.}

\item{output_dir}{Where output should be saved to.}
}
\description{
Saves statistics on confidence and data for each participant to .csv files.
}
