% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_orphans.R
\name{remove_orphans}
\alias{remove_orphans}
\title{Clean up your filtered dataset.}
\usage{
remove_orphans(dat, id_num, output_dir)
}
\arguments{
\item{dat}{Dataframe.}

\item{id_num}{Participant identifier. Must be stored like so, e.g., "OpenFace_01234_Mother".
Defaults to the first column in a dataframe.}

\item{output_dir}{Where you want the output to be saved to.}
}
\description{
Try not to leave any orphans behind. Specific to dyadic data.
}
