% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smush_rows.R
\name{smush_rows}
\alias{smush_rows}
\title{Average rows by a specified amount of time.}
\usage{
smush_rows(dat, id_num, smush)
}
\arguments{
\item{dat}{Dataframe.}

\item{id_num}{Your participant identifier. Defaults to first column in the dataframe.}

\item{smush}{To what extent do you want to average rows together?
For example, if you specify \code{10}, then each chunk of 10 rows will be averaged together.
Keep in mind that each row is 1/3 of a second. Defaults to 10.}
}
\description{
Average rows by a specified amount of time, e.g., 5 frames, 10 frame
Be aware that this will remove the following variables: frame, timestamp, confidence, and success.
}
