% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_list.R
\name{load_list}
\alias{load_list}
\title{Read in a directory of OpenFace files and do stuff.}
\usage{
load_list(path)
}
\arguments{
\item{path}{Path to your directory of \code{.csv} files after running OpenFace.}
}
\description{
This will load all OpenFace \code{.csv} files as a \code{list}.
Make sure to save it to an object or it won't work.
}
