% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_stack.R
\name{long_stack}
\alias{long_stack}
\title{Merge dataframes from a list into one long dataframe}
\usage{
long_stack(yourlist)
}
\arguments{
\item{yourlist}{List, a list of dataframes.}
}
\description{
Merge all OpenFace dataframes from a list into a single long-form dataframe.
This will take a very long time, so it comes with a progress bar.
Your list of dataframes must already be read into R.
}
