% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fake.R
\name{make_fake}
\alias{make_fake}
\title{Create fake OpenFace data.}
\usage{
make_fake(path, n, loop, seed)
}
\arguments{
\item{path}{Character, path to where you want the files to go.}

\item{n}{Numeric, number of rows for each generated \code{.csv}}

\item{loop}{Numeric, total number of individual \code{.csv} files to generate.}

\item{seed}{Numeric, random seed.}
}
\description{
Each \code{.csv} file will have a random 15-digit ID number with the prefix "fake_".
Excluding the informational (e.g., \code{faceid}, \code{timestamp}) and the Action Unit data, the values don't make sense.
}
