% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_small.R
\name{make_small}
\alias{make_small}
\title{Make OpenFace \code{.csv} files narrower}
\usage{
make_small(bigpath, newpath)
}
\arguments{
\item{bigpath}{Character, path to where all your big files currently live.}

\item{newpath}{Character, path to where you want to write out the new smaller files.}
}
\description{
Subset only relevant columns from OpenFace data: the first five columns and the Action Units.
}
