% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_dyad.R
\name{pivot_dyad}
\alias{pivot_dyad}
\title{Widen data by dyad member.}
\usage{
pivot_dyad(dat, id_num, smushed, output_dir)
}
\arguments{
\item{dat}{Dataframe.}

\item{id_num}{Participant identifier. Defaults to first column in dataframe.}

\item{smushed}{Binary, TRUE/FALSE. Has your data been smushed with \code{smush_rows()}? Required.}

\item{output_dir}{Where you want your dataframes to be saved to.}
}
\description{
Widen dataframe by dyad member.
At this point, you should have merged your data using \code{load_list} and \verb{long stack},
filtered using relevant criteria using \code{filter_rows},
removed orphans with \code{remove_orphans}, and smushed with \code{smush_rows}.
}
\details{
For dyadic data, this is the last step in the workflow.

This will output two dataframes: 1) an extra long dataframe by dyad member, and 2)
a slightly wider dataframe where each dyad members has their own Action Unit column.
}
